Homework 6 -Packages refactor from capital letters
