+++ src/com/mobile/store/brand/Xiaomi.java
@@ -1,4 +1,4 @@
-package com.mobile.store.Brand;
+package com.mobile.store.brand;
 
 import com.mobile.store.Smartphone;
